move to using scoped css modules for each component
