+++ gatsby-browser.js
@@ -4,4 +4,4 @@
  * See: https://www.gatsbyjs.org/docs/browser-apis/
  */
 
-import './src/styles/app.scss';
+import './src/styles/global.scss';
