+++ package.json
@@ -61,7 +61,7 @@
     "prettier:write": "npm run prettier -- --write",
     "lint:js": "eslint --ext .js,.jsx,.ts,.tsx ./src",
     "lint:js:fix": "npm run lint:js -- --fix",
-    "lint:css": "stylelint ./src/styles",
+    "lint:css": "stylelint '**/*.scss'",
     "lint:css:fix": "npm run lint:css -- --fix",
     "prod:build": "gatsby build",
     "prod:serve": "gatsby serve",
