+++ src/components/primitives/Header/Header.tsx
@@ -1,14 +1,15 @@
 import { Link } from 'gatsby';
 import PropTypes from 'prop-types';
 import React from 'react';
-import { homePath } from '../../routing/urlGenerator';
+import { homePath } from '../../../routing/urlGenerator';
+import styles from './header.module.scss';
 
 type Props = {
     siteTitle: string;
 };
 
 const Header = ({ siteTitle }: Props) => (
-    <header className="header">
+    <header className={styles.header}>
         <div className="container">
             <h1>
                 <Link to={homePath}>{siteTitle}</Link>
