+++ src/components/primitives/Layout.tsx
@@ -8,7 +8,7 @@
 import React, { ReactNode } from 'react';
 import PropTypes from 'prop-types';
 import { useStaticQuery, graphql } from 'gatsby';
-import Header from './Header';
+import Header from './Header/Header';
 
 type Props = {
     children: ReactNode;
