--- src/styles/components/_header.scss
@@ -1,14 +1,0 @@
-.header {
-    margin-bottom: 2rem;
-    padding: 1rem;
-    background-color: #663;
-
-    h1 {
-        margin-bottom: 0;
-    }
-
-    a {
-        color: #fff;
-        text-decoration: none;
-    }
-}
