+++ src/styles/tools/_media-queries.scss
@@ -1,0 +1,3 @@
+// import this inside scss module to use sass-mq breakpoints
+@import '~@freshheads/maatcss/settings/breakpoints';
+@import 'sass-mq';
